Use AST in project_analysis.py instead of inspect module.
The difference here is, instead of actually loading modules and their
subsequent dependencies (which can be slow or cause errors), we now work
directly of the AST of source files and inject a little bit of logic
that would normally be done during module execution to match things up
(e.g. references to variables and their actual declaration).