Now using typeshed stubs where possible.
Processing time of typing/ without:
~11s.

With typeshed:
~1s.

Still longer than it needs to be realistically - it's not that much code -
but I think the issue is that it's still going through a lot of code via
dependencies ala numpy, parso, etc.

It'll be fun to profile the shit out of this.